+++ etna/analysis/outliers/rolling_statistics.py
@@ -314,7 +314,7 @@ def mad_method(
     if mad_scale <= 0:
         raise ValueError("Scaling parameter must be positive!")
 
-    window = series[indices]
+    window = series.iloc[indices]
 
     if trend or seasonality:
         if stl_params is None: